requirements.txt
python-dotenv
crewai
pydantic
PyYAML